Fix smoke test to run under pbuilder
pbuilder provides a hook to run autopkgtest after build,
but doesn't like it very much when there are stray processes
around when it's cleaning up. The result is that our smoke
test, which passes when run using autopkgtest-virt-lxc, gets
stuck when run using pbuilder.

Fix it to clean up properly after itself. As a side effect
of the rewrite, dbus-x11 is no longer needed.