Upgrade libvirt-daemon-system from Suggests to Depends
While it's possible for libvirt-dbus to work even when the package
is not installed, that will result in crippled functionality as
the user will only be able to connect to the autostarted session
daemon and not to the (often more interesting) system daemon.

Given the above, it's reasonable to play it safe and drag in a few
extra files in order to make sure users won't be frustrated because
of missing functionality.

We can always reconsider this decision down the line.

Gbp-Dch: Short