Add dbus-session-bus to Depends
Similarly to what happens with libvirt-daemon-system, lacking
this package does not mean that libvirt-dbus won't work at all,
but functionality will nonetheless be severely impacted because
users won't be able to connect to the session daemon.

Make life easier for users by dragging in support for the D-Bus
session bus explicitly. We can always revisit this decision at
a later time.

Gbp-Dch: Short