patches: Backport recent fixes
The following patches contain fixes that have already been
accepted upstream but are not part of any release yet:

  backport/Do-not-use-gulong-for-storing-uint64.patch
  backport/Handle-APIs-expecting-unsigned-long-correctly.patch
  backport/Handle-APIs-returning-unsigned-long-correctly.patch
  backport/tests-vcpupininfo-use-pytest.skip.patch
  backport/tests-skip-CPU-pinning-test-on-libvirt-6.6.0.patch
  backport/src-fix-double-free-of-virDomain-object.patch

In particular, known FTBFS with libvirt 6.6.0 and on 32bit
architectures such as armhf, as well as random test suite failures
that show up from time to time, are addressed.

Thanks to Christian Ehrhardt for pointing out the issue, working
to get it fixed upstream and taking a first stab at this patch.

Closes: #971171