Switch to manual user management
We're already doing part of the job manually because dh-sysuser
doesn't support declaring a list of additional groups that the
newly-created user should be a part of.

In addition to that, it turns out that dh-sysuser is buggy and
doesn't clean up properly when the package is purged.

Luckily our needs in terms of user management are fairly basic,
so implementing the necessary logic ourselves is not complicated.

https://bugs.debian.org/1000478

Gbp-Dch: Ignore