rules: Explicitly disable deep systemd integration
If the build system detects that the host is running systemd,
it will generate and install an alternative configuration which
will attempt to start libvirt-dbus as a systemd service instead
of a regular process when it's not already running.

As long as Debian still supports using alternative init systems,
we should avoid configuring libvirt-dbus in a way that can only
work with systemd.