Only add libvirtdbus user to libvirt group on Ubuntu
Debian already installs a policykit rule to achieve the intended
result, and it would be confusing for the user if privileged
libvirt access was granted through two separate mechanisms.

We can still keep the code that deals with group membership
around, since doing so doesn't introduce a significant
maintenance burden for us and makes life easier for the Ubuntu
developers, but at the same time we don't want Debian users to
be affected by it so let's run it conditionally.

Gbp-Dch: Ignore