Depend on polkitd instead of policykit-1
The polkit package has been made more modular, and libvirt-dbus
only needs the daemon part.

Note that we explicitly ask for support for the legacy pkla
format to be present because, while Debian now contains a version
of polkit recent enough to support the JavaScript-based rules
that upstream includes, Ubuntu still doesn't.

Closes: #1025579