Don't depend on a specific dbus implementation
libvirt-dbus works just fine with alternative dbus implementations
such as dbus-broker, and we can express this by depending on the
appropriate virtual packages.

The upstream test suite, which is run as part of each package
build, and the autopkgtest suite, however, call the dbus-daemon
command explicitly, so we need to ensure that we have that
specific implementation around in those contexts.

We could make things a bit nicer there by using the dbus-daemon
package, which doesn't drag in quite as much stuff as the dbus
package does, but that would break backportability to Ubuntu
20.04 so we're leaving things as they are for now.

See: https://lists.debian.org/debian-devel/2022/11/msg00114.html